+++ src/flows/user-program-management/helpers/questions.ts
@@ -1,3 +1,5 @@
+import Entropy from "@entropyxyz/sdk";
+
 export const addQuestions = [
   {
     type: "input",
@@ -28,3 +30,11 @@ export const getProgramPointerInput = [
     message: "Enter the program pointer you wish to remove:",
   },
 ]
+
+export const verifyingKeyQuestion = (entropy: Entropy) => [{
+  type: 'list',
+  name: 'verifyingKey',
+  message: 'Select the key to proceeed',
+  choices: entropy.keyring.accounts.registration.verifyingKeys,
+  default: entropy.keyring.accounts.registration.verifyingKeys[0]
+}]